refactor: merge type and damage class icons into single shared component
