--- src/components/TypeIcon.astro
@@ -1,37 +1,0 @@
----
-import { cva } from 'cva'
-
-import { TYPES, type TypeKey } from '@/lib/domain/types'
-
-const variants = cva({
-  base: 'flex flex-row items-center justify-center',
-  variants: {
-    size: {
-      small: 'size-5 rounded-xs',
-      medium: 'size-6 rounded-sm',
-      large: 'size-8 rounded-md',
-    },
-  },
-})
-
-interface Props {
-  variant: string
-  size?: 'small' | 'medium' | 'large'
-  class?: string
-}
-
-const { variant, size = 'large', class: className } = Astro.props
-const name = TYPES[variant as TypeKey]
-const imageUrl = `${variant}.png`
-const dimensions = { small: 20, medium: 28, large: 32 }
----
-
-<div title={name} class:list={['relative', `bg-${variant}`, variants({ size }), className]}>
-  <img
-    src={imageUrl}
-    alt={name}
-    width={dimensions[size]}
-    height={dimensions[size]}
-    class="absolute object-contain"
-  />
-</div>
