+++ src/components/details/evolution/EvolutionChip.astro
@@ -15,7 +15,7 @@ import Swords from '@lucide/astro/icons/swords'
 import Users from '@lucide/astro/icons/users'
 import Venus from '@lucide/astro/icons/venus'
 
-import TypeIcon from '@/components/TypeIcon.astro'
+import { TypeIcon } from '@/components/icons'
 import type { Chip, ChipIconName } from '@/lib/utils/evolution-conditions'
 
 interface Props {
