+++ src/components/details/evolution/EvolutionMonster.astro
@@ -1,7 +1,7 @@
 ---
 import type { Pokemon, PokemonSpecies } from 'pokedex-promise-v2'
 
-import TypeIcon from '@/components/TypeIcon.astro'
+import { TypeIcon } from '@/components/icons'
 import { toHref } from '@/lib/utils/path'
 import type { TypeKey } from '@/lib/domain/types'
 import { getTranslation } from '@/lib/utils/pokeapi-helpers'
