+++ src/styles/globals.css
@@ -1,11 +1,8 @@
 @import url('https://fonts.googleapis.com/css2?family=Geist:ital,wght@0,100..900;1,100..900&family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&family=JetBrains+Mono:ital,wght@0,100..800;1,100..800&family=Sofia+Sans:ital,wght@0,1..1000;1,1..1000&display=swap');
 @import 'tailwindcss';
 
-/* Type colors are applied dynamically: bg-<slug>. */
-/* Classes are built at runtime so Tailwind so we safelist the full sets here. */
-@source inline(
-  'bg-{normal,fighting,flying,poison,ground,rock,bug,ghost,steel,fire,water,grass,electric,psychic,ice,dragon,dark,fairy}'
-);
+@source inline('bg-{normal,fighting,flying,poison,ground,rock,bug,ghost,steel,fire,water,grass,electric,psychic,ice,dragon,dark,fairy}');
+@source inline('bg-{physical,special,status}');
 
 @view-transition {
   navigation: auto;
@@ -26,14 +23,6 @@
   --color-physical: #eb5628;
   --color-special: #375ab2;
   --color-status: #828282;
-
-  --color-physical-light: #f29173;
-  --color-special-light: #7d94cd;
-  --color-status-light: #aeaeae;
-
-  --color-physical-dark: #99381a;
-  --color-special-dark: #243b74;
-  --color-status-dark: #555555;
 
   --color-normal: #9fa19f;
   --color-fighting: #ff8000;
